fixed altered files
