+++ lesson_15/tdd/tdd_app/src/test/java/com/codedifferently/lesson15/EmployeeManagerTest.java
@@ -1,0 +1,49 @@
+package com.codedifferently.lesson15;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+public class EmployeeManagerTest {
+
+  EmployeeManager employeeManager;
+
+  @BeforeEach
+  void setUp() {
+    employeeManager = new EmployeeManager();
+  }
+
+  @Test
+  void testaddEmployee() {
+    employeeManager.addEmployee("1");
+  }
+
+  @Test
+  void testgetEployee() {
+    // assert
+  }
+
+  @Test
+  void testupdateEmployee() {
+    // add stuff
+  }
+
+  @Test
+  void testremoveEmployee() {
+    // add stuff
+  }
+
+  @Test
+  void testassertEmployeeInCollection() {
+    // add stuff
+    // throw new is needed
+  }
+
+  @Test
+  void testgetEmployeeCount() {}
+
+  @Test
+  public void testGetGreeting() {
+    // Act
+    Lesson15.main(null);
+  }
+}
