+++ lesson_15/tdd/tdd_app/src/test/java/com/codedifferently/lesson15/Lesson15Test.java
@@ -1,48 +1,16 @@
 package com.codedifferently.lesson15;
 
-import org.junit.jupiter.api.BeforeEach;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import org.junit.jupiter.api.Test;
 
 class Lesson15Test {
 
-  EmployeeManager employeeManager;
-
-  @BeforeEach
-  void setUp() {
-    employeeManager = new EmployeeManager();
+  @Test
+  public void testLesson15() {
+    assertThat(new Lesson15()).isNotNull();
   }
 
-  @Test
-  void testaddEmployee() {
-    employeeManager.addEmployee("1");
-  }
-  
-  @Test
-  void testgetEployee() {
-    //assert
-  }
-
-  @Test
-  void testupdateEmployee() {
-    //add stuff
-  }
-
-  @Test
-  void testremoveEmployee() {
-    //add stuff
-  }
-
-  @Test
-  void testassertEmployeeInCollection() {
-    //add stuff
-    //throw new is needed
-  }
-
-  @Test
-  void testgetEmployeeCount() {
-    
-  }
-  
   @Test
   public void testGetGreeting() {
     // Act
